Create Todo.py
In this program, I defined a list called "tasks" to store the tasks. I also defined four functions: "add task" to add a task to the list, "view tasks" to display all the tasks in the list, "remove task" to remove a task from the list, and "main" to implement the program logic. The user can choose from four options to interact with the program.